+++ src/app/utils.js
@@ -1,0 +1,9 @@
+import Application from './application';
+
+export default {
+	anchorNavigation: (e) => {
+		e.preventDefault();
+		const { href } = e.target;
+		Application.navigate(href);
+	},
+};